Store fuzzified regions in layer so we could apply top surface fuzzify later
